+++ R/zzz.R
@@ -12,6 +12,8 @@
   make_decision_tree_rpart()
   make_decision_tree_partykit()
 
+  make_null_model_censored()
+
   make_proportional_hazards_survival()
   make_proportional_hazards_glmnet()